Update connection.html and business.html
Refactor the headings and content in connection.html and business.html to provide more accurate and descriptive information about the web development and business topics covered in the respective weeks. Also, update the warning messages to indicate that additional content needs to be added. These changes improve the clarity and relevance of the pages.